+++ mock/volume.go
@@ -8,6 +8,7 @@ import (
 	context "context"
 	io "io"
 	reflect "reflect"
+	time "time"
 
 	gomock "github.com/golang/mock/gomock"
 )
@@ -36,17 +37,17 @@ func (m *Volume) EXPECT() *VolumeMockRecorder {
 }
 
 // CreateFile mocks base method.
-func (m *Volume) CreateFile(arg0 context.Context, arg1 string, arg2 io.ReadCloser, arg3 int) error {
+func (m *Volume) CreateFile(arg0 context.Context, arg1 string, arg2 io.ReadCloser, arg3 int, arg4 time.Duration, arg5 time.Time) error {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "CreateFile", arg0, arg1, arg2, arg3)
+	ret := m.ctrl.Call(m, "CreateFile", arg0, arg1, arg2, arg3, arg4, arg5)
 	ret0, _ := ret[0].(error)
 	return ret0
 }
 
 // CreateFile indicates an expected call of CreateFile.
-func (mr *VolumeMockRecorder) CreateFile(arg0, arg1, arg2, arg3 interface{}) *gomock.Call {
+func (mr *VolumeMockRecorder) CreateFile(arg0, arg1, arg2, arg3, arg4, arg5 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CreateFile", reflect.TypeOf((*Volume)(nil).CreateFile), arg0, arg1, arg2, arg3)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CreateFile", reflect.TypeOf((*Volume)(nil).CreateFile), arg0, arg1, arg2, arg3, arg4, arg5)
 }
 
 // DeleteFile mocks base method.