+++ uow/unit_of_work.go
@@ -12,6 +12,7 @@ import (
 	"github.com/spf13/afero"
 	"github.com/xescugc/rebost/file"
 	"github.com/xescugc/rebost/idxkey"
+	"github.com/xescugc/rebost/idxttl"
 	"github.com/xescugc/rebost/idxvolume"
 	"github.com/xescugc/rebost/replica"
 	"github.com/xescugc/rebost/state"
@@ -36,6 +37,7 @@ const (
 type UnitOfWork interface {
 	Files() file.Repository
 	IDXKeys() idxkey.Repository
+	IDXTTLs() idxttl.Repository
 	IDXVolumes() idxvolume.Repository
 	Fs() afero.Fs
 	Replicas() replica.Repository